% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPL_data.R
\name{EPL_dat}
\alias{EPL_dat}
\title{League of Legends data}
\format{A list with four elements.
\item A data frame with 494 rows and 15 variables:
  \describe{
    \item{price}{price, in US dollars}
    \item{carat}{weight of the diamond, in carats}
  }
\item A vector of teams.}
\source{
\url{http://www.diamondse.info/}
}
\description{
A dataset containing the LCK statistics for the summer 2018 season.
}
