% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_CFA_approx.R
\name{L_CFA_approx}
\alias{L_CFA_approx}
\title{Compute normal copula likelihood}
\usage{
L_CFA_approx(obs, mu, sd, Gamma, logl = T, sd_train)
}
\arguments{
\item{obs}{The data point.}

\item{mu}{The estimated mean parameter.}

\item{sd}{The estimated standard deviation.}

\item{Gamma}{The estimated covariance matrix.}

\item{logl}{Should the output be log-likelihood?}

\item{sd_train}{The standard deviation of the original data.}
}
\value{
The (log-)likelihood.
}
\description{
This function computes the approximated likelihood of a Gaussian copula with
normal marginals. Integration in the calculation of the exact
likelihood can be computationally expensive, especially in high dimensions,
therefore this function provides us with an approximation of the integral.
}
