% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_CPFA_approx.R
\name{L_CPFA_approx}
\alias{L_CPFA_approx}
\title{Compute approximate copula likelihood with Poisson marginals}
\usage{
L_CPFA_approx(obs, Theta, Gamma, logl = T)
}
\arguments{
\item{obs}{vector. The data point.}

\item{Theta}{vector. The estimated Poisson parameter for this observation.}

\item{Gamma}{matrix. The estimated correlation matrix.}

\item{logl}{logical. Should the output be log-likelihood?}
}
\value{
The (log-)likelihood.
}
\description{
This function computes the approximated likelihood of a Gaussian copula with
Poisson marginals. Integration in the calculation of the exact
likelihood can be computationally expensive, especially in high dimensions,
therefore this function provides us with an approximation of the integral.
}
