% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_NBFA_approx.R
\name{L_NBFA_approx}
\alias{L_NBFA_approx}
\title{Compute NB copula likelihood}
\usage{
L_NBFA_approx(obs, size, mu, Gamma, logl = T)
}
\arguments{
\item{obs}{The data point.}

\item{size}{The estimated size parameter.}

\item{mu}{The estimated mean parameter.}

\item{Gamma}{The estimated covariance matrix.}

\item{logl}{Should the output be log-likelihood?}
}
\value{
The (log-)likelihood.
}
\description{
This function computes the approximated likelihood of a Gaussian copula with
NB marginals. Integration in the calculation of the exact
likelihood can be computationally expensive, especially in high dimensions,
therefore this function provides us with an approximation of the integral.
}
