% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_cl.R
\name{likelihood_cl}
\alias{likelihood_cl}
\title{Calculate the likelihood for the FA and gFA models.}
\usage{
likelihood_cl(test, pred, tr_fa, logl = TRUE, integrate = FALSE,
  truncate = FALSE)
}
\arguments{
\item{test}{data frame. Test data set for evaluation.}

\item{pred}{data frame or matrix. The provided predictions.}

\item{tr_fa}{Test dataset (for evaluation).}

\item{L_intg}{logical. Should the likelihood be integrated across
suitable borders? Defaults to FALSE, which uses an approximation with
MVN density.
Special care needs to be taken when integrating, as the likelihood is
likely to be 0 in higher dimensions due to machine precision.}
}
\value{
numeric vector. The (log) likelihood for each test observation.
}
\description{
This function calculates the MVN likelihood.
}
\seealso{
\code{\link{my_dmvn}} \code{\link{my_omxMnor}}
}
