% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dmvn.R
\name{my_dmvn}
\alias{my_dmvn}
\title{Calculate the density}
\usage{
my_dmvn(my_vec, l, logl)
}
\arguments{
\item{my_vec}{numeric vector. First \code{l} values represent the
observations, the next \code{l} values represent the predictions (means),
and
the last \code{l}^2 values represent the estimated residual covariance
matrix.}

\item{l}{numeric. Number of dimensions.}

\item{logl}{logical. Calculate log density? Defaults to true.}
}
\value{
numeric vector. The (log) density.
}
\description{
This function calculates the MVN density.
}
