% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_omxMnor.R
\name{my_omxMnor}
\alias{my_omxMnor}
\title{Calculate the integrated density}
\usage{
my_omxMnor(my_vec, l, logl)
}
\arguments{
\item{my_vec}{numeric vector. First \code{l}^2 values represent the
estimated residual covariance, the next \code{l} values represent
the predictions (means), the next \code{l} values the lower bounds for
integration, and the last \code{l} values the upper bounds for integration.}

\item{l}{numeric. Number of dimensions.}

\item{logl}{logical. Calculate log density? Defaults to true.}
}
\value{
numeric vector. The (log) density.
}
\description{
This function calculates the MVN density by integrating along
corresponding bounds.
}
