% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_CFA.R
\name{pred_CFA}
\alias{pred_CFA}
\title{Predict from CFA model}
\usage{
pred_CFA(tr_fa, gp_test, test, logl = TRUE)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_CFA}.}

\item{gp_test}{vector. Groups for each observation in the test set.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the CFA
model \code{train_CFA}.
}
\seealso{
\code{\link{train_CFA}}
}
