% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_CNBFA1.R
\name{pred_CNBFA1}
\alias{pred_CNBFA1}
\title{Predict from CNBFA1 model}
\usage{
pred_CNBFA1(tr_fa, gp_test, test, L_intg = F, L_trnc = F, scale_dat = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_CNBFA1}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the CNBFA1
model \code{train_CNBFA1}.
}
\seealso{
\code{\link{train_CNBFA1}}
}
