% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_CPFA.R
\name{pred_CPFA}
\alias{pred_CPFA}
\title{Predict from CPFA model}
\usage{
pred_CPFA(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_CPFA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the CPFA
model \code{train_CPFA}.
}
\seealso{
\code{\link{train_CPFA}}
}
