% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_FA.R
\name{pred_FA}
\alias{pred_FA}
\title{Predict from FA model}
\usage{
pred_FA(tr_fa, gp_test, test, logl = T, L_intg = F)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_FA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}

\item{L_intg}{logical. Should the likelihood be integrated across
suitable borders? Defaults to FALSE, which uses an approximation with
MVN density.
Special care needs to be taken when integrating, as the likelihood is
likely to be 0 in higher dimensions due to machine precision.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{log_lik}{A numeric vector of (log-)likelihoods for all observations
  in the test set.}
}
\description{
This function predicts based on the estimation of the FA
model \code{train_FA}.
}
\seealso{
\code{\link{train_FA}}
}
