% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NBFA2.R
\name{test_NBFA2}
\alias{test_NBFA2}
\title{Predict from NBFA2 model}
\usage{
test_NBFA2(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_NBFA2}.}

\item{gp_test}{A vector of groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log) likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the NBFA2
model \code{train_NBFA2}.
}
\seealso{
\code{\link{train_NBFA2}}
}
