% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_PFA.R
\name{test_PFA}
\alias{test_PFA}
\title{Predict from PFA model}
\usage{
test_PFA(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_PFA}.}

\item{gp_test}{A vector of groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log) likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the PFA
model \code{train_PFA}.
}
\seealso{
\code{\link{train_PFA}}
}
