% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_CFA.R
\name{train_CFA}
\alias{train_CFA}
\title{Train a copula FA model}
\usage{
train_CFA(train, gp_train, nfac = 2, ...)
}
\arguments{
\item{train}{data frame. The training data set, consisting of counts.
Columns represent variables, rows represent observations.}

\item{gp_train}{vector. Groups for each observation.}

\item{nfac}{numeric. The number of factors.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter,
chains).}
}
\value{
A list.
\item{train}{Training data set.}
\item{gp_train}{A vector of groups for each observation.}
\item{loadings}{Aggregated loadings.}
\item{scores}{Aggregated factor scores.}
\item{noise}{Aggregated residual covariance matrix.}
\item{noise_org}{Aggregated residual covariance matrix, scaled by the
variance of the original data.}
\item{Sigma}{Aggregated correlation matrix.}
\item{stan_mod}{A list with two elements.
Both are objects of S4 class \code{stanfit}. First is the marginal
distribution estimation, second is the copula.}
}
\description{
This function trains a copula factor analysis model in Stan.
}
\details{
The transformations are made separately, between the modeling of
marginals and the correlation.
}
