% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_iPoiss.R
\name{train_iPoiss}
\alias{train_iPoiss}
\title{Train an independent Poisson model}
\usage{
train_iPoiss(train, gp_train)
}
\arguments{
\item{train}{data frame. The training data set, consisting of counts.
Columns represent variables, rows represent observations.}

\item{gp_train}{vector. Groups for each observation.}
}
\value{
list with three elements. First is the Poisson parameters for
  each variable-group combination. Second is the train data set. Third is
  the group vector.
}
\description{
This function trains an independent Poisson model for multivariate
count estimation.
}
