% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_NBFA2.R
\name{train_NBFA2}
\alias{train_NBFA2}
\title{Train a negative binomial factor analysis model}
\usage{
train_NBFA2(train, gp_train, nfac = 2, ...)
}
\arguments{
\item{train}{data frame. The training data set, consisting of counts.
Columns represent variables, rows represent observations.}

\item{gp_train}{vector. Groups for each observation.}

\item{nfac}{numeric. The number of factors.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter,
chains).}
}
\value{
A list.
\item{train}{Training data set.}
\item{gp_train}{A vector of groups for each observation.}
\item{loadings}{Aggregated loadings.}
\item{scores}{Aggregated factor scores.}
\item{mu}{Aggregated mean parameters.}
\item{probs}{Aggregated probability parameters.}
\item{phi}{Aggregated size parameters.}
\item{stan_mod}{An object of S4 class \code{stanfit}.}
}
\description{
This function trains a negative binomial factor analysis model in Stan.
Each variable has its own scaling parameter.
}
