% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_PFA.R
\name{train_PFA}
\alias{train_PFA}
\title{Train a Poisson factor analysis model}
\usage{
train_PFA(train, gp_train, nfac = 2, ...)
}
\arguments{
\item{train}{data frame. The training data set, consisting of counts.
Columns represent variables, rows represent observations.}

\item{gp_train}{vector. Groups for each observation.}

\item{nfac}{numeric. The number of factors.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter,
chains).}
}
\value{
A list.
\item{train}{Training data set.}
\item{gp_train}{A vector of groups for each observation.}
\item{loadings}{Aggregated loadings.}
\item{scores}{Aggregated factor scores.}
\item{stan_mod}{An object of S4 class \code{stanfit}.}
}
\description{
This function trains a Poisson factor analysis model in Stan.
}
