% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulafa-package.R
\docType{package}
\name{copulafa-package}
\alias{copulafa-package}
\alias{copulafa}
\title{The 'copulafa' package.}
\description{
The package contains several factor analysis methods, with
emphasis to count data. The main methods rely on copulas to estimate the
residual covariance, not explained by the latent structure. Additionally, the
package contains four data sets for empirical evaluation and one baseline
model. The methods use RStan for inference, the only exception is the
iPoiss method where we use conjugate posterior.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
