% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_iPoiss.R
\name{pred_iPoiss}
\alias{pred_iPoiss}
\title{Predict from an independent Poisson model}
\usage{
pred_iPoiss(tr_fa, gp_test, test)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_iPoiss()}.}

\item{gp_test}{vector. Groups for each observation in the test set.}

\item{test}{data frame. Test data set for evaluation.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{log_lik}{A numeric vector of log-likelihoods for all observations
  in the test set.}
}
\description{
This function predicts based on the estimation of the independent Poisson
model \code{train_iPoiss}.
}
\seealso{
\code{\link{train_iPoiss}}
}
