% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_iPoiss.R
\name{train_iPoiss}
\alias{train_iPoiss}
\title{Train an independent Poisson model}
\usage{
train_iPoiss(train, gp_train)
}
\arguments{
\item{train}{data frame. The training data set, consisting of counts.
Columns represent variables, rows represent observations.}

\item{gp_train}{vector. Groups for each observation.}
}
\value{
A list.
  \item{params}{A matrix of Poisson parameters for each group-variable
  combination. Rows represent groups, columns represent variables.}
  \item{train}{Train data set.}
  \item{gp_train}{Groups of the train data set.}
}
\description{
This function trains an independent Poisson model for multivariate
count estimation.
}
\seealso{
\code{\link{pred_iPoiss}}
}
